HU2: refactor ids a Long y GET de usuario por id

+++ applications/app-service/src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS users (
-    id SERIAL PRIMARY KEY,
-    id_req_res INTEGER NOT NULL UNIQUE,
+    id BIGSERIAL PRIMARY KEY,
+    id_req_res BIGINT NOT NULL UNIQUE,
     email VARCHAR(255) NOT NULL,
     first_name VARCHAR(255) NOT NULL,
     last_name VARCHAR(255) NOT NULL
