+++ domain/model/src/main/java/co/com/retowebflux/model/user/gateways/UserNoSqlRepository.java
@@ -4,7 +4,7 @@ import co.com.retowebflux.model.user.User;
 import reactor.core.publisher.Mono;
 
 public interface UserNoSqlRepository {
-    Mono<User> findById(Integer id);
+    Mono<User> findById(Long id);
 
     Mono<User> save(User user);
 }
