+++ infrastructure/driven-adapters/r2dbc-postgresql/src/main/java/co/com/retowebflux/r2dbc/UserReactiveRepository.java
@@ -3,6 +3,6 @@ package co.com.retowebflux.r2dbc;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Mono;
 
-public interface UserReactiveRepository extends ReactiveCrudRepository<UserEntity, Integer> {
-    Mono<UserEntity> findByIdReqRes(Integer idReqRes);
+public interface UserReactiveRepository extends ReactiveCrudRepository<UserEntity, Long> {
+    Mono<UserEntity> findByIdReqRes(Long idReqRes);
 }
